Move TestMemoryFork, TestPromptManager, TestTurnState, and TestUtilities into internals/testing
