+++ js/packages/teams-ai/src/Application.spec.ts
@@ -17,9 +17,9 @@ import {
 import { Application, ConversationUpdateEvents, MessageReactionEvents, TeamsMessageEvents } from './Application';
 import { AdaptiveCardsOptions } from './AdaptiveCards';
 import { AIOptions } from './AI';
+import { createTestConversationUpdate, createTestInvoke } from './internals/testing/TestUtilities';
 import { TaskModulesOptions } from './TaskModules';
 import { TurnState } from './TurnState';
-import { createTestConversationUpdate, createTestInvoke } from './internals';
 import { TestPlanner } from './planners/TestPlanner';
 import { TeamsAdapter } from './TeamsAdapter';
