+++ js/packages/teams-ai/src/MessageExtensions.spec.ts
@@ -1,7 +1,7 @@
 import sinon from 'sinon';
 import { strict as assert } from 'assert';
 import { Application, Query } from './Application';
-import { createTestInvoke } from './internals/TestUtilities';
+import { createTestInvoke } from './internals/testing/TestUtilities';
 import { MessageExtensions, MessageExtensionsInvokeNames } from './MessageExtensions';
 import {
     Channels,
