+++ js/packages/teams-ai/src/internals/index.ts
@@ -9,5 +9,4 @@
 export * from './AzureOpenAIClient';
 export * from './Colorize';
 export * from './OpenAIClient';
-export * from './TestUtilities';
 export * from './types';
