+++ js/packages/teams-ai/src/internals/testing/TestMemoryFork.ts
@@ -1,4 +1,4 @@
-import { Memory } from './MemoryFork';
+import { Memory } from '../../MemoryFork';
 
 /**
  * @private
