+++ js/packages/teams-ai/src/prompts/TemplateSection.spec.ts
@@ -1,9 +1,9 @@
 import { strict as assert } from 'assert';
 import { TemplateSection } from './TemplateSection';
 import { TestAdapter } from 'botbuilder';
-import { TestPromptManager } from './TestPromptManager';
+import { TestPromptManager } from '../internals/testing/TestPromptManager';
+import { TestTurnState } from '../internals/testing/TestTurnState';
 import { GPT3Tokenizer } from '../tokenizers';
-import { TestTurnState } from '../TestTurnState';
 
 describe('TemplateSection', () => {
     const adapter = new TestAdapter();
