+++ js/packages/teams-ai/src/prompts/UserMessage.spec.ts
@@ -1,9 +1,9 @@
 import { strict as assert } from 'assert';
-import { UserMessage } from './UserMessage';
 import { TestAdapter } from 'botbuilder';
-import { TestPromptManager } from './TestPromptManager';
+import { TestPromptManager } from '../internals/testing/TestPromptManager';
+import { TestTurnState } from '../internals/testing/TestTurnState';
 import { GPT3Tokenizer } from '../tokenizers';
-import { TestTurnState } from '../TestTurnState';
+import { UserMessage } from './UserMessage';
 
 describe('UserMessage', () => {
     const adapter = new TestAdapter();
