+++ js/packages/teams-ai/src/prompts/index.ts
@@ -22,7 +22,6 @@ export * from './PromptSectionBase';
 export * from './PromptTemplate';
 export * from './SystemMessage';
 export * from './TemplateSection';
-export * from './TestPromptManager';
 export * from './TextSection';
 export * from './UserInputMessage';
 export * from './UserMessage';
