+++ js/packages/teams-ai/src/validators/JSONResponseValidator.spec.ts
@@ -1,11 +1,11 @@
 import { strict as assert } from 'assert';
 import * as sinon from 'sinon';
+import { TestAdapter } from 'botbuilder';
 
+import { TestTurnState } from '../internals/testing/TestTurnState';
+import { GPT3Tokenizer } from '../tokenizers';
 import { JSONResponseValidator } from './JSONResponseValidator';
 import { Schema, ValidationError } from 'jsonschema';
-import { TestAdapter } from 'botbuilder';
-import { GPT3Tokenizer } from '../tokenizers';
-import { TestTurnState } from '../TestTurnState';
 
 describe('JSONResponseValidator', () => {
     const adapter = new TestAdapter();
